Update Rewrites.hs
I was struggling with the same issue for myMinimumBy and myMaximumBy and then it struck me. It specifically states that it returns the last value the comparison returned GT for (in case of maximum) so if all comparisons return LT you should end up at the last value and just return that, just like the proposed example inputs/outputs suggests. That's how I got the idea to do a foldl and start with the last value of the list. And the other way around for minimum.